Minor changes to allow running flense-nw from a browser
Unfortunately running browser REPL, remote devtools for React tools and cljs-devtools is not easily possible at this point with current version of NS.JS.

Fortunately I was able to do that all when running Chrome Canary with flense-nw loaded as a standard web-page.

Added some supporting files for this more of operation and short explanation.